Separate Python code from specfem2d.cpp
