+++ src/specfem2d.cpp
@@ -1,56 +1,4 @@
-#include "compute/interface.hpp"
-// #include "coupled_interface/interface.hpp"
-// #include "domain/interface.hpp"
-#include "IO/interface.hpp"
-#include "kokkos_abstractions.h"
-#include "mesh/mesh.hpp"
-#include "parameter_parser/interface.hpp"
-#include "receiver/interface.hpp"
-#include "solver/solver.hpp"
-#include "source/interface.hpp"
-#include "specfem_mpi/interface.hpp"
-#include "specfem_setup.hpp"
-#include "timescheme/timescheme.hpp"
-#include "yaml-cpp/yaml.h"
-#include <Kokkos_Core.hpp>
-#include <boost/program_options.hpp>
-#include <chrono>
-#include <ctime>
-#include <memory>
-#include <stdexcept>
-#include <string>
-#include <vector>
-#ifdef SPECFEMPP_BINDING_PYTHON
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-#define STRINGIFY(x) #x
-#define MACRO_STRINGIFY(x) STRINGIFY(x)
-#include "periodic_tasks/check_signal.hpp"
-#endif
-// Specfem2d driver
-
-std::string
-print_end_message(std::chrono::time_point<std::chrono::system_clock> start_time,
-                  std::chrono::duration<double> solver_time) {
-  std::ostringstream message;
-  // current date/time based on current system
-  const auto now = std::chrono::system_clock::now();
-
-  std::time_t c_now = std::chrono::system_clock::to_time_t(now);
-
-  std::chrono::duration<double> diff = now - start_time;
-
-  message << "\n================================================\n"
-          << "             Finished simulation\n"
-          << "================================================\n\n"
-          << "Total simulation time : " << diff.count() << " secs\n"
-          << "Total solver time (time loop) : " << solver_time.count()
-          << " secs\n"
-          << "Simulation end time : " << ctime(&c_now)
-          << "------------------------------------------------\n";
-
-  return message.str();
-}
+#include "execute.hpp"
 
 boost::program_options::options_description define_args() {
   namespace po = boost::program_options;
@@ -89,307 +37,6 @@ int parse_args(int argc, char **argv,
   return 1;
 }
 
-void execute(const YAML::Node &parameter_dict, const YAML::Node &default_dict,
-             specfem::MPI::MPI *mpi) {
-
-  // --------------------------------------------------------------
-  //                    Read parameter file
-  // --------------------------------------------------------------
-  auto start_time = std::chrono::system_clock::now();
-  specfem::runtime_configuration::setup setup(parameter_dict, default_dict);
-  const auto database_filename = setup.get_databases();
-  mpi->cout(setup.print_header(start_time));
-
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Read mesh and materials
-  // --------------------------------------------------------------
-  const auto quadrature = setup.instantiate_quadrature();
-  const auto mesh = specfem::IO::read_mesh(database_filename, mpi);
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Read Sources and Receivers
-  // --------------------------------------------------------------
-  const int nsteps = setup.get_nsteps();
-  const specfem::simulation::type simulation_type = setup.get_simulation_type();
-  auto [sources, t0] =
-      specfem::IO::read_sources(setup.get_sources(), nsteps, setup.get_t0(),
-                                setup.get_dt(), simulation_type);
-  setup.update_t0(t0); // Update t0 in case it was changed
-
-  const auto stations_node = setup.get_stations();
-  const auto angle = setup.get_receiver_angle();
-  auto receivers = specfem::IO::read_receivers(stations_node, angle);
-
-  mpi->cout("Source Information:");
-  mpi->cout("-------------------------------");
-  if (mpi->main_proc()) {
-    std::cout << "Number of sources : " << sources.size() << "\n" << std::endl;
-  }
-
-  for (auto &source : sources) {
-    mpi->cout(source->print());
-  }
-
-  mpi->cout("Receiver Information:");
-  mpi->cout("-------------------------------");
-
-  if (mpi->main_proc()) {
-    std::cout << "Number of receivers : " << receivers.size() << "\n"
-              << std::endl;
-  }
-
-  for (auto &receiver : receivers) {
-    mpi->cout(receiver->print());
-  }
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Instantiate Timescheme
-  // --------------------------------------------------------------
-  const auto time_scheme = setup.instantiate_timescheme();
-  if (mpi->main_proc())
-    std::cout << *time_scheme << std::endl;
-
-  const int max_seismogram_time_step = time_scheme->get_max_seismogram_step();
-
-  const int nstep_between_samples = time_scheme->get_nstep_between_samples();
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Generate Assembly
-  // --------------------------------------------------------------
-  mpi->cout("Generating assembly:");
-  mpi->cout("-------------------------------");
-  const type_real dt = setup.get_dt();
-  specfem::compute::assembly assembly(
-      mesh, quadrature, sources, receivers, setup.get_seismogram_types(),
-      setup.get_t0(), dt, nsteps, max_seismogram_time_step,
-      nstep_between_samples, setup.get_simulation_type(),
-      setup.instantiate_property_reader());
-  time_scheme->link_assembly(assembly);
-
-  // --------------------------------------------------------------
-  //                Read or Write properties
-  // --------------------------------------------------------------
-  const auto property_writer = setup.instantiate_property_writer();
-  if (property_writer) {
-    mpi->cout("Writing model files:");
-    mpi->cout("-------------------------------");
-
-    property_writer->write(assembly);
-    return;
-  }
-
-  // const auto property_reader = setup.instantiate_property_reader();
-  // if (property_reader) {
-  //   mpi->cout("Reading model files:");
-  //   mpi->cout("-------------------------------");
-
-  //   property_reader->read(assembly);
-  // }
-
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Read wavefields
-  // --------------------------------------------------------------
-
-  const auto wavefield_reader = setup.instantiate_wavefield_reader();
-  if (wavefield_reader) {
-    mpi->cout("Reading wavefield files:");
-    mpi->cout("-------------------------------");
-
-    wavefield_reader->read(assembly);
-    // Transfer the buffer field to device
-    assembly.fields.buffer.copy_to_device();
-  }
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Instantiate plotter
-  // --------------------------------------------------------------
-  std::vector<std::shared_ptr<specfem::periodic_tasks::periodic_task> > tasks;
-  const auto wavefield_plotter = setup.instantiate_wavefield_plotter(assembly);
-  tasks.push_back(wavefield_plotter);
-
-#ifdef SPECFEMPP_BINDING_PYTHON
-  const auto signal_task =
-      std::make_shared<specfem::periodic_tasks::check_signal>(10);
-  tasks.push_back(signal_task);
-#endif
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Instantiate Solver
-  // --------------------------------------------------------------
-  std::shared_ptr<specfem::solver::solver> solver =
-      setup.instantiate_solver<5>(dt, assembly, time_scheme, tasks);
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Execute Solver
-  // --------------------------------------------------------------
-  // Time the solver
-  mpi->cout("Executing time loop:");
-  mpi->cout("-------------------------------");
-
-  const auto solver_start_time = std::chrono::system_clock::now();
-  solver->run();
-  const auto solver_end_time = std::chrono::system_clock::now();
-
-  std::chrono::duration<double> solver_time =
-      solver_end_time - solver_start_time;
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Write Seismograms
-  // --------------------------------------------------------------
-  const auto seismogram_writer = setup.instantiate_seismogram_writer();
-  if (seismogram_writer) {
-    mpi->cout("Writing seismogram files:");
-    mpi->cout("-------------------------------");
-
-    seismogram_writer->write(assembly);
-  }
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                  Write Forward Wavefields
-  // --------------------------------------------------------------
-  const auto wavefield_writer = setup.instantiate_wavefield_writer();
-  if (wavefield_writer) {
-    mpi->cout("Writing wavefield files:");
-    mpi->cout("-------------------------------");
-
-    wavefield_writer->write(assembly);
-  }
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                Write Kernels
-  // --------------------------------------------------------------
-  const auto kernel_writer = setup.instantiate_kernel_writer();
-  if (kernel_writer) {
-    mpi->cout("Writing kernel files:");
-    mpi->cout("-------------------------------");
-
-    kernel_writer->write(assembly);
-  }
-  // --------------------------------------------------------------
-
-  // --------------------------------------------------------------
-  //                   Print End Message
-  // --------------------------------------------------------------
-  mpi->cout(print_end_message(start_time, solver_time));
-  // --------------------------------------------------------------
-
-  return;
-}
-
-#ifdef SPECFEMPP_BINDING_PYTHON
-
-namespace py = pybind11;
-
-// global MPI variable for Python
-specfem::MPI::MPI *_py_mpi = NULL;
-
-bool _initialize(py::list py_argv) {
-  if (_py_mpi != NULL) {
-    return false;
-  }
-  // parse argc and argv from Python
-  int argc = py_argv.size();
-  char **argv = new char *[argc + 1];
-
-  for (size_t i = 0; i < argc; i++) {
-    std::string str =
-        py_argv[i].cast<std::string>(); // Convert Python string to std::string
-    argv[i] =
-        new char[str.length() + 1]; // Allocate memory for each C-style string
-    std::strcpy(argv[i], str.c_str()); // Copy the string content
-  }
-
-  // Null-terminate argv following the specification
-  argv[argc] = nullptr;
-  // Initialize MPI
-  _py_mpi = new specfem::MPI::MPI(&argc, &argv);
-  // Initialize Kokkos
-  Kokkos::initialize(argc, argv);
-
-  // free argv
-  for (int i = 0; i < argc; i++) {
-    delete[] argv[i]; // Free each individual string
-  }
-
-  delete[] argv;
-
-  return true;
-}
-
-bool _execute(const std::string &parameter_string,
-              const std::string &default_string) {
-  if (_py_mpi == NULL) {
-    return false;
-  }
-  const YAML::Node parameter_dict = YAML::Load(parameter_string);
-  const YAML::Node default_dict = YAML::Load(default_string);
-  execute(parameter_dict, default_dict, _py_mpi);
-  return true;
-}
-
-bool _finalize() {
-  if (_py_mpi != NULL) {
-    // Finalize Kokkos
-    Kokkos::finalize();
-    // Finalize MPI
-    delete _py_mpi;
-    _py_mpi = NULL;
-    return true;
-  }
-  return false;
-}
-
-PYBIND11_MODULE(_core, m) {
-    m.doc() = R"pbdoc(
-        SPECfem++ core module
-        -----------------------
-
-        .. currentmodule:: specfempp
-
-        .. autosummary::
-           :toctree: _generate
-
-           _run
-    )pbdoc";
-
-    m.def("_initialize", &_initialize, R"pbdoc(
-        Initialize SPECFEM++.
-    )pbdoc");
-
-    m.def("_execute", &_execute, R"pbdoc(
-        Execute the main SPECFEM++ workflow.
-    )pbdoc");
-
-    m.def("_finalize", &_finalize, R"pbdoc(
-        Finalize SPECFEM++.
-    )pbdoc");
-
-    m.attr("_default_file_path") = __default_file__;
-
-#ifdef VERSION_INFO
-    m.attr("__version__") = MACRO_STRINGIFY(VERSION_INFO);
-#else
-    m.attr("__version__") = "dev";
-#endif
-}
-
-#else // SPECFEMPP_BINDING_PYTHON
-
 int main(int argc, char **argv) {
   // Initialize MPI
   specfem::MPI::MPI *mpi = new specfem::MPI::MPI(&argc, &argv);
@@ -403,7 +50,9 @@ int main(int argc, char **argv) {
       const std::string default_file = vm["default_file"].as<std::string>();
       const YAML::Node parameter_dict = YAML::LoadFile(parameters_file);
       const YAML::Node default_dict = YAML::LoadFile(default_file);
-      execute(parameter_dict, default_dict, mpi);
+      std::vector<std::shared_ptr<specfem::periodic_tasks::periodic_task> >
+          tasks;
+      execute(parameter_dict, default_dict, tasks, mpi);
     }
   }
   // Finalize Kokkos
@@ -412,5 +61,3 @@ int main(int argc, char **argv) {
   delete mpi;
   return 0;
 }
-
-#endif // SPECFEMPP_BINDING_PYTHON
